Implemented #25, changed fitness to double

+++ JNF_NEAT/body.h
@@ -8,7 +8,7 @@ class IBody
 
 		virtual void Reset() = 0;
 		virtual void Update(const std::vector<float>& networkOutputs) = 0;
-		virtual int GetFitness() const = 0;
+		virtual double GetFitness() const = 0;
 
 		virtual std::vector<float> ProvideNetworkWithInputs() const = 0;
 };
